importer-msgraph-metadata: ensure leading service name is trimmed from operation names when using a recognised verb
